+++ packages/pagekit/blog/src/Controller/SiteController.php
@@ -85,18 +85,20 @@ class SiteController
         $site = App::module('system/site');
         $feed = App::feed()->create($type ?: $this->blog->config('feed.type'), [
             'title' => $site->config('title'),
-            'link' => App::url('@blog', [], true),
+            'link' => App::url('@blog', [], 0),
             'description' => $site->config('description'),
             'element' => ['language', $locale],
-            'selfLink' => App::url('@blog/feed', [], true)
+            'selfLink' => App::url('@blog/feed', [], 0)
         ]);
 
         if ($last = Post::where(['status = ?', 'date < ?'], [Post::STATUS_PUBLISHED, new \DateTime])->limit(1)->orderBy('modified', 'DESC')->first()) {
             $feed->setDate($last->modified);
         }
 
-        foreach (Post::where(['status = ?', 'date < ?'], [Post::STATUS_PUBLISHED, new \DateTime])->related('user')->limit($this->blog->config('feed.limit'))->orderBy('date', 'DESC')->get() as $post) {
-            $url = App::url('@blog/id', ['id' => $post->id], true);
+        foreach (Post::where(['status = ?', 'date < ?'], [Post::STATUS_PUBLISHED, new \DateTime])->where(function ($query) {
+            return $query->where('roles IS NULL')->whereInSet('roles', App::user()->roles, false, 'OR');
+        })->related('user')->limit($this->blog->config('feed.limit'))->orderBy('date', 'DESC')->get() as $post) {
+            $url = App::url('@blog/id', ['id' => $post->id], 0);
             $feed->addItem(
                 $feed->createItem([
                     'title' => $post->title,
@@ -130,10 +132,23 @@ class SiteController
 
         $user = App::user();
 
+        $description = $post->get('meta.og:description');
+        if (!$description) {
+            $description = strip_tags($post->excerpt ?: $post->content);
+            $description = rtrim(mb_substr($description, 0, 150), " \t\n\r\0\x0B.,") . '...';
+        }
+
         return [
             '$view' => [
                 'title' => __($post->title),
-                'name' => 'blog/post.php'
+                'name' => 'blog/post.php',
+                'og:type' => 'article',
+                'article:published_time' => $post->date->format(\DateTime::ATOM),
+                'article:modified_time' => $post->modified->format(\DateTime::ATOM),
+                'article:author' => $post->user->name,
+                'og:title' => $post->get('meta.og:title') ?: $post->title,
+                'og:description' => $description,
+                'og:image' =>  $post->get('image.src') ? App::url()->getStatic($post->get('image.src'), [], 0) : false
             ],
             '$comments' => [
                 'config' => [